LLM/customer_service_system/requirements.txt
requests>=2.31.0
Pillow>=10.0.0
python-dotenv>=1.0.0 